+++ luscii_patient_actions_sdk_ios/lib/luscii_patient_actions_sdk_ios.dart
@@ -53,9 +53,9 @@ class LusciiPatientActionsSdkIOS extends LusciiPatientActionsSdkPlatform {
   }
 
   @override
-  Future<List<dynamic>> getSelfcareActions() async {
+  Future<List<dynamic>> getSelfCareActions() async {
     final actions = await methodChannel.invokeMethod<List<dynamic>>(
-      'getSelfcareActions',
+      'getSelfCareActions',
     );
 
     if (actions is! List) {