+++ luscii_patient_actions_sdk_android/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.1+1] - 2026-01-23
+
+### Fixed
+- Fixed bug where `isPlanned`, `isSelfCare` and `isExtra` were null.
+- Renamed `getSelfcareActions` to `getSelfCareActions`.
+
 ## [0.8.0+2] - 2026-01-21
 
 ### Fixed